MAINT: Renamed new_analysis_model.py to analysis_model.py
