+++ force_wfmanager/model/analysis_model.py
