+++ force_wfmanager/ui/review/i_data_view.py
@@ -1,6 +1,6 @@
 from traits.api import Interface, Instance, Bool
 
-from force_wfmanager.model.new_analysis_model import AnalysisModel
+from force_wfmanager.model.analysis_model import AnalysisModel
 
 
 class IDataView(Interface):
