+++ force_wfmanager/ui/review/tests/test_data_view_pane.py
@@ -1,7 +1,7 @@
 import mock
 import unittest
 
-from force_wfmanager.model.new_analysis_model import AnalysisModel
+from force_wfmanager.model.analysis_model import AnalysisModel
 from force_wfmanager.ui.review.data_view_pane import DataViewPane
 from force_wfmanager.ui.review.plot import Plot
