+++ force_wfmanager/ui/review/tests/test_results_pane.py
@@ -1,6 +1,6 @@
 import unittest
 
-from force_wfmanager.model.new_analysis_model import AnalysisModel
+from force_wfmanager.model.analysis_model import AnalysisModel
 from force_wfmanager.ui.review.results_pane import ResultsPane
